vzmemcheck: fix overflow in show total ram
On systems with more than 8G of RAM we have an overflow when multiplying
(long) physpages to (long) pagesize. Cast it to (unsigned long long) to fix.

This should fix the followning uglyness in vzmemcheck -V output:

Summary:    135.15 1547.91 1684.90 1684.90 17381.22 2509.13 17381.22 49931.91
            984.00  984.00 17592186044317.00 8090.00 8090.00 8090.00 8090.00 8090.00
                           ^^^^^^^^^^^^^^^^^

Reported-by: Jakub Jankowski <shasta+ovz@toxcorp.com>

+++ src/lib/util.c
@@ -367,7 +367,7 @@ int get_mem(unsigned long long *mem)
 	}
 	if ((pagesize = get_pagesize()) < 0)
 		return -1;
-	*mem = pages * pagesize;
+	*mem = (unsigned long long) pages * pagesize;
 	return 0;
 }
